pyproject.toml
[tool.poetry]
name = "trident-sync"
version = "1.1.2"
description = "三叉戟，二次开发项目同步升级工具，Secondary development repo sync and upgrade CLI"
authors = ["xiaojunnuo <xiaojunnuo@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "lib" }, { include = "cli.py" }, { include = "pyproject.toml" }]

[tool.poetry.dependencies]
python = ">=3.8,<4"
loguru = "^0.6.0"
docopt = "^0.6.2"
gitpython = "^3.1.30"
pyyaml = "^6.0"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-html = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.scripts]
trident = "cli:cli"